+++ lib/features/post_list/presentation/view_models/pagination_manager.dart
@@ -1,6 +1,6 @@
+import 'package:sodong_app/features/post/domain/use_case/post_service.dart';
 import 'package:sodong_app/features/post_list/domain/models/region.dart';
 import 'package:sodong_app/features/post_list/domain/models/town_life_post.dart';
-import 'package:sodong_app/features/post_list/domain/usecase/post_service.dart';
 
 /// 페이지네이션 관리 클래스
 ///
